* develop: (28 commits)
  Mise en place de JalonActivity commun pour les activités d'un cours
  Boite de dépot lien relatif OK
  Ajout Suppression Ok sans les relatives items
  Afficher / Masquer OK
  Icone + roue dentelée ok sur liste documents
  Listing des documents enseignants OK
  Ajout des documents enseignants dans boite de dépôts OK
  Test sur les dates sur la page de gestion de la grille d'évaluation
  Changement des options de la boîte de dépots
  Test OK + annuler et modifier le coeff d'un critère de la grille d'évaluation
  Test sur mode_etudiant et non affichage des boutons pour les étudiants
  Interface de création, modification, suppression OK.
  Suppression d'un critère OK
  Edition d'un critère OK
  Affichage du formulaire d'édition d'un critère
  Affichage de toutes les infos d'un critère
  Visualisation de la grille d'évaluation OK
  Uniformisation des noms suite
  Renommage pour uniformisation, suppression de JalonXXX et passage du français à l'anglais
  Mise en place des vérifications sur les dates en mode évaluation par les pairs
  ...

# Conflicts:
#	jalon.content/jalon/content/content/jalonfolder.py

+++ jalon.content/jalon/content/skins/jalon_content_course/course_delete_item_script.py
@@ -10,7 +10,12 @@
 
 form = context.REQUEST.form
 
-context.deleteCourseMapItem(form["item_id"], None)
-context.deleteCourseActuality(form["item_id"])
+redirection = context.absolute_url()
+if context.meta_type == "JalonCours":
+    context.deleteCourseMapItem(form["item_id"], None)
+    context.deleteCourseActuality(form["item_id"])
+else:
+    context.detachDocument(form["item_id"])
+    redirection = "%s?tab=documents" % context.absolute_url()
 
-context.REQUEST.RESPONSE.redirect(context.absolute_url())
+context.REQUEST.RESPONSE.redirect(redirection)
